+++ src/PRo3D.Core/Surface.fs
@@ -188,6 +188,8 @@ module SurfaceIntersection =
                                 let trafo = 
                                     if surf.transformation.flipZ then 
                                         Trafo3d.Scale(1.0, 1.0, -1.0).Forward * (fullTrafo.Forward * surf.preTransform.Forward)
+                                    else if surf.transformation.isSketchFab then
+                                        Sg.switchYZTrafo.Forward
                                     else
                                         (fullTrafo.Forward * surf.preTransform.Forward)
 
@@ -202,6 +204,8 @@ module SurfaceIntersection =
                                 let backward = 
                                     if surf.transformation.flipZ then
                                         Trafo3d.Scale(1.0, 1.0, -1.0).Backward * surf.preTransform.Backward * fullTrafo.Backward
+                                    else if surf.transformation.isSketchFab then
+                                        Sg.switchYZTrafo.Backward
                                     else
                                         surf.preTransform.Backward * fullTrafo.Backward